Moved PermissionListener to permissions module
