+++ collect_app/src/main/java/org/odk/collect/android/permissions/PermissionsProvider.kt
@@ -16,8 +16,8 @@ import com.karumi.dexter.listener.PermissionGrantedResponse
 import com.karumi.dexter.listener.PermissionRequest
 import com.karumi.dexter.listener.multi.MultiplePermissionsListener
 import org.odk.collect.android.R
-import org.odk.collect.android.listeners.PermissionListener
 import org.odk.collect.androidshared.ui.DialogFragmentUtils.showIfNotShowing
+import org.odk.collect.permissions.PermissionListener
 import org.odk.collect.permissions.PermissionsChecker
 import timber.log.Timber
