+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/FormMetadataPreferencesFragment.java
@@ -17,11 +17,11 @@ import androidx.preference.Preference;
 import org.jetbrains.annotations.NotNull;
 import org.odk.collect.android.R;
 import org.odk.collect.android.injection.DaggerUtils;
-import org.odk.collect.android.listeners.PermissionListener;
 import org.odk.collect.android.logic.PropertyManager;
 import org.odk.collect.android.permissions.PermissionsProvider;
 import org.odk.collect.androidshared.ui.ToastUtils;
 import org.odk.collect.shared.strings.Validator;
+import org.odk.collect.permissions.PermissionListener;
 
 import javax.inject.Inject;
