+++ collect_app/src/test/java/org/odk/collect/android/fakes/FakePermissionsProvider.kt
@@ -2,8 +2,8 @@ package org.odk.collect.android.fakes
 
 import android.app.Activity
 import androidx.test.platform.app.InstrumentationRegistry
-import org.odk.collect.android.listeners.PermissionListener
 import org.odk.collect.android.permissions.PermissionsProvider
+import org.odk.collect.permissions.PermissionListener
 import org.odk.collect.permissions.PermissionsChecker
 
 /**
